+++ src/compile
@@ -1,13 +1,1 @@
-rm ATG/*.java
-
-rm ../out/src/*.class
-
-rm ../out/SemanticScope/*class
-
-rm ../out/SemanticClasses/*.class
-
-rm ../out/SyntacticTree/*.class
-
-java -jar ../libraries/Coco.jar ATG/OpenHugs.ATG 
-
-javac -d ../out SyntacticTree/*.java *.java SemanticClasses/*.java SemanticScope/*.java
+javac -d ../out/ main/java/SemanticClasses/*.java main/java/SemanticScope/*.java main/java/SyntacticTree/*.java main/java/*.java
