+++ src/main/java/DataType.java
@@ -1,4 +1,4 @@
-package src;
+package main.java;
 
 public enum DataType {
     NUMERO,
