+++ src/main/java/SemanticClasses/WhileStatement.java
@@ -1,24 +1,24 @@
+package main.java.SemanticClasses;
 
-package main.java.SemanticTree;
+import java.util.HashMap;
 
-import main.java.SemanticScope.DataType;
+import main.java.*;
 
-import java.util.HashMap; //Diccionario :v
-
-public class IfStatement implements Statement {
+public class WhileStatement implements Statement{
 
     private HashMap<String, DataType> contenido;
     private String operador;
     private DataType tipoExp = null;
+    StatementType stType;
 
-    public IfStatement() {
-        this.type = "SI";
-
+    public WhileStatement(){
+        this.stType = StatementType.CICLO;
     }
+
     public void setOperador(String operador){ //Se puede meter todo tipo de operadore pero solo asigna cuando encuentre el indicado
         if (operador.equals("<=") || operador.equals(">=")
-         || operador.equals("<")  || operador.equals(">")
-         || operador.equals("==") || operador.equals("!=") )
+                || operador.equals("<")  || operador.equals(">")
+                || operador.equals("==") || operador.equals("!=") )
 
         {
             this.operador = operador;
@@ -50,7 +50,7 @@ public class IfStatement implements Statement {
 
                     resultado = true;
                 }
-            break;
+                break;
 
             case LOGICO:
                 if(     contenido.containsValue(DataType.LOGICO) &&
@@ -60,7 +60,7 @@ public class IfStatement implements Statement {
 
                     resultado = true;
                 }
-            break;
+                break;
 
             case PALABRA:
                 if(         contenido.containsValue(DataType.PALABRA)
@@ -69,14 +69,13 @@ public class IfStatement implements Statement {
                         && (operador.equals("==") || operador.equals("!="))){//Si es por palabra solo acepta que se comparen strings)
                     resultado = true;
                 }
-            break;
+                break;
         }
 
         return  resultado;
-
     }
-
-
-
-
+    
+    public StatementType getType(){
+        return this.stType;
+    }
 }
