--- src/main/java/SemanticScope/DataType.java
@@ -1,9 +1,0 @@
-package main.java.SemanticScope;
-
-public enum DataType {
-    NUMERO,
-    LOGICO,
-    PALABRA,
-    NONE,
-    OPERATOR
-}
