--- src/main/java/SemanticTree/FunctionStatement.java
@@ -1,8 +1,0 @@
-package main.java.SemanticTree;
-
-public class FunctionStatement implements Statement{
-    
-    public FunctionStatement(){
-        this.type = "FUNCION";
-    }
-}
