--- src/main/java/SemanticTree/Statement.java
@@ -1,6 +1,0 @@
-package main.java.SemanticTree;
-
-public interface Statement{
-
-    
-}
