+++ src/main/java/StatementType.java
@@ -1,4 +1,4 @@
-package src;
+package main.java;
 
 public enum StatementType {
     SI,
